+++ Data/SBV/Provers/Prover.hs
@@ -429,16 +429,16 @@ instance (SymWord a, Provable p) => Provable (SBV a -> p) where
 
 -- SFunArrays (memory, functional representation), only supported universally for the time being
 instance (HasKind a, HasKind b, Provable p) => Provable (SArray a b -> p) where
-  forAll_       k = declNewSArray (\t -> "array_" ++ show t) >>= \a -> forAll_   $ k a
-  forAll (s:ss) k = declNewSArray (const s)                  >>= \a -> forAll ss $ k a
+  forAll_       k = newArray_  >>= \a -> forAll_   $ k a
+  forAll (s:ss) k = newArray s >>= \a -> forAll ss $ k a
   forAll []     k = forAll_ k
   forSome_      _ = error "SBV.forSome.SFunArray: Existential arrays are not currently supported."
   forSome _     _ = error "SBV.forSome.SFunArray: Existential arrays are not currently supported."
 
 -- SArrays (memory, SMT-Lib notion of arrays), only supported universally for the time being
 instance (HasKind a, HasKind b, Provable p) => Provable (SFunArray a b -> p) where
-  forAll_       k = declNewSFunArray (\t -> "funArray_" ++ show t) >>= \a -> forAll_   $ k a
-  forAll (s:ss) k = declNewSFunArray (const s)                     >>= \a -> forAll ss $ k a
+  forAll_       k = newArray_  >>= \a -> forAll_   $ k a
+  forAll (s:ss) k = newArray s >>= \a -> forAll ss $ k a
   forAll []     k = forAll_ k
   forSome_      _ = error "SBV.forSome.SArray: Existential arrays are not currently supported."
   forSome _     _ = error "SBV.forSome.SArray: Existential arrays are not currently supported."